refactor(game): peel the career-startup options out of the game commands
Everything that turns the player's choices on the new-career screen into a
clock: `RawStartupOptions` and the validated `StartupOptions` it becomes,
`StartPhase`, the year floor, the date and age arithmetic, and how much past
history to generate behind the opening season.

Moves to `game/startup.rs` with all fifteen of its tests. `mod.rs` 3475 → 3332.

This is the cleanest seam in the file, and the tests are why. All fifteen call
only things that came with them — `normalize_startup_options`,
`start_date_for_year`, `age_on_date`, `preseason_*`,
`apply_generated_past_history` — with one exception:
`age_on_date_uses_world_snapshot_date_over_startup_phase` also needs
`game_clock_for_world`, which belongs to the world-loading cluster and has not
moved yet. It imports it from the parent rather than following it, because what
the test is actually about is `age_on_date` preferring a snapshot date over the
startup phase.

Three end-to-end tests stay in `mod.rs` and still construct a `StartupOptions`
directly, so the struct, its fields and `DEFAULT_GENERATED_HISTORY_DEPTH_YEARS`
are `pub(super)` rather than private. Everything else moved keeps the reach it
had; only `start_phase_for_game` and `StartPhase` are re-exported `pub(crate)`,
for `mcp_server/tools_impl/game.rs`.

Leaf-for-leaf test list unchanged: 216 lib tests before, 216 after (225 with
`--features mcp`, which is what CI builds).
